made recommended changes
