+++ hw8/tasks/firsttask.ts
@@ -17,7 +17,7 @@ interface Address {
     geo: Geo
 }
 
-interface User {
+export interface User {
     id: number;
     name: string,
     username: string,
@@ -29,7 +29,7 @@ interface User {
 }
 
 
-class UserInfo {
+export class UserInfo {
 
     public user: User;
 
@@ -51,21 +51,3 @@ class UserInfo {
 }
 
 
-async function logUsers (): Promise<void> {
-    try {
-        const response = await fetch('https://jsonplaceholder.typicode.com/users');
-        const users: User[] = await response.json();
-        mutateUser(users[0]);
-    } catch (error) {
-        console.error('Error fetching data:', error);
-    }
-}
-
-function mutateUser(user: User): void {
-    const userInfo = new UserInfo(user);
-    console.log(userInfo.getAddress());
-    console.log(userInfo.getCompany());
-    console.log(userInfo.getUserName());
-}
-
-logUsers();
