--- hw8/tasks/index.js
@@ -1,18 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var abstraction_1 = require("./abstraction");
-var dog = new abstraction_1.Dog('Leo', 10, 'Goldendoodle');
-console.log(dog);
-dog.bark();
-dog.eat(1);
-dog.play();
-dog.sleep();
-console.log(dog.energy);
-var cat = new abstraction_1.Cat('Charles', 10, false);
-console.log(cat);
-console.log(cat.declawed);
-cat.eat(2);
-cat.meow();
-cat.play();
-cat.sleep();
-console.log(cat.energy);
