+++ hw8/tasks/index.ts
@@ -1,4 +1,5 @@
-import { Dog, Cat } from "./abstraction";
+import { Dog, Cat } from './abstraction';
+import {UserInfo, User} from './firsttask';
 
 const dog = new Dog('Leo', 10, 'Goldendoodle');
 
@@ -22,5 +23,24 @@ cat.play();
 cat.sleep();
 
 console.log(cat.energy);
+
+async function logUsers (): Promise<void> {
+    try {
+        const response = await fetch('https://jsonplaceholder.typicode.com/users');
+        const users: User[] = await response.json();
+        mutateUser(users[0]);
+    } catch (error) {
+        console.error('Error fetching data:', error);
+    }
+}
+
+function mutateUser(user: User): void {
+    const userInfo = new UserInfo(user);
+    console.log(userInfo.getAddress());
+    console.log(userInfo.getCompany());
+    console.log(userInfo.getUserName());
+}
+
+logUsers();
